+++ src/main/java/frc/robot/Robot.java
@@ -9,6 +9,7 @@ package frc.robot;
 
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.command.Scheduler;
+import frc.robot.subsystems.Rider;
 
 /**
  * The VM is configured to automatically run this class, and to call the
@@ -92,6 +93,8 @@ public class Robot extends TimedRobot {
   @Override
   public void teleopPeriodic() {
     Scheduler.getInstance().run();
+
+    
   }
 
   /**
