+++ src/main/java/frc/robot/commands/Rider/AngleOutward.java
@@ -5,16 +5,16 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.Rider;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.subsystems.Rider;
 
-public class AngleInwards extends Command {
-
+public class AngleOutward extends Command {
+  
   private Rider rider = Rider.getInstance();
 
-  public AngleInwards() {
+  public AngleOutward() {
     // Use requires() here to declare subsystem dependencies
     requires(rider);
   }
@@ -27,7 +27,7 @@ public class AngleInwards extends Command {
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    rider.angleButtonInward();
+    rider.angleButtonOutward();
   }
 
   // Make this return true when this Command no longer needs to run execute()
@@ -39,6 +39,9 @@ public class AngleInwards extends Command {
   // Called once after isFinished returns true
   @Override
   protected void end() {
+
+    rider.stopAngleMotor();
+
   }
 
   // Called when another command which requires one or more of the same
