+++ src/main/java/frc/robot/commands/Rider/IntakeIn.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.Rider;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.subsystems.Rider;
@@ -40,6 +40,9 @@ public class IntakeIn extends Command {
   // Called once after isFinished returns true
   @Override
   protected void end() {
+
+    rider.stopIntake();    
+
   }
 
   // Called when another command which requires one or more of the same
